+++ dplib/models/field/constraints/string.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from typing import Optional
+
+from .collection import CollectionConstraints
+
+
+class StringConstraints(CollectionConstraints):
+    pattern: Optional[str] = None
