+++ dplib/models/field/datatypes/number.py
@@ -1,0 +1,31 @@
+from __future__ import annotations
+
+from typing import Literal, Optional
+
+import pydantic
+
+from ..constraints import EnumerableConstraints
+from .base import BaseField
+
+
+class NumberField(BaseField):
+    type: Literal["integer"] = "integer"
+    format: Optional[Literal["default"]] = None
+    constraints: EnumerableConstraints = pydantic.Field(
+        default_factory=EnumerableConstraints
+    )
+
+    bareNumber: Optional[bool] = None
+    """
+    If false leading and trailing non numbers will be removed for integer/number fields
+    """
+
+    groupChar: Optional[str] = None
+    """
+    String whose value is used to group digits for integer/number fields
+    """
+
+    decimalChar: Optional[str] = None
+    """
+    String whose value is used to represent a decimal point for number fields
+    """
